Split snitch-view script
This separation is due to a distinction in Kubernetes contexts, one for
the watched cluster and another for Snitch.

In the <setup_snitch_view.sh> script remains the functions which create
the Secret with the generated kubeconfig and a Cluster instance. A new
script called <gen_snitch_view_kubeconfig.sh> has been added, being
responsible for creating a service account and configure RBAC for it.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>