Implement discovery of cluster source
Changes:
	- Add functions to discover provider, flavor and region;
	- Include new discovery functions in the <ClusterDiscoverer>
	  interface;

In order to fetch this data, Snitch performs label maching against the
cluster nodes' metadata, keeping a collection of label to cluster source
mappings in the <discovery> package.

For the provider and flavor, only the 1st node is checked. As for the
region, Snitch considers the one with the biggest amount of nodes.

A function called <ClusterSource> extracts both the provider and flavor
data, wherefrom the <Provider> and <Flavor> functions obtain their
result. The later 2 are yet to be used, but have been implemented to
remain complaint with the <ClusterDiscoverer> interface.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>