Add Github actions
This commit brings two new Github actions. The 1st is triggered on every
push to the repository and on pull requests to the main branch, building
the project and running unit tests over it. The 2nd executes whenever
version tags are pushed, building and pushing Docker images for
Inspect's core and server components.

Docker-run's "-t" option has been removed to permit execution on non TTY
devices.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>