Create worker structure
Changes:
	- Add <config> package;
	- Add parser for Popeye Json reports;

The <config> package is mainly responsible for transporting cluster data
to the <ClusterIssue> instances which will be created as a result of the
worker's processing, as well as to customize some of the worker's
behavior. For now, this customization is limited to the location
wherefrom the "done" file will be expected.

A blueprint of Popeye's Json output is used to unmarshal the report,
which is afterwards transformed into a slice of <ClusterIssueSpec>.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>