Add ownerref to <ClusterIssue> instances
The <report.Parse> function now uses the Job information from the config
to set each <ClusterIssue> instance as child of the Job that triggered
it. Furthermore, the Job name suffix is appended to the <ClusterIssue>
instance name. This naming convention adds another layer of organization
when relating each issue instance to their Job.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>