Update handler unit tests
Remove a test case covering the flow in which the server'd ignore an
issue when it comes from a scan with at least one plugin failure, given
this behavior is no longer present. The argument for scans with failure
has been removed as well.

Also fix the UI Chart svcaccount template.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>