Adapt Hml workflow for open Zora
Part of the Github Hml workflow fetches the latest tag from Zora's Chart
repository. Given the recent repository split still uses the same Chart
name, it becomes possible for a repository workflow to fetch a version
created on the other.

This commit changes the workflow to only consider the versions matching
the pattern:
	<verion_tag>-open

Thus, future Chart versions for this repository ought to have the
"-open" suffix.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>