Halt plugin Cronjob upon unready Cluster
Zora schedules plugin scans through Cronjobs, which are created on its
1st execution.

In case something went wrong with the connected cluster, the
<ClusterScan> controller would emit an error, but the Cronjob were still
trying to execute, even though the scan is bound to fail.

This commit makes the scan reconciler suspend the plugin Cronjob
whenever the condition <Ready> from the scan's <Cluster> instance is
false.

The <Suspend> status field from <ClusterScan> will now reflect this
halt as well.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>