Use a unique prefix for the default sample manifest name
The generated <Cluster> manifest is persisted to a file whose name is
controlled by the environment variable <SAMPLE_MANIFEST_NAME>. This
variable used to default to "cluster_sample.yaml", what'd cause
consecutive executions to overwrite the file in case the variable is not
customized.

This commit introduces a prefix for the default name, according to the
K8s context already fetched by the script.

Furthermore, the output order has also been changed, where the lines
informing the user about the names of the generated files appear first.

Signed-off-by: Carlos Torres <carlos.torres@getupcloud.com>